modal, remove-task, nopage restructured
