+++ src/components/Modal/modal.scss
@@ -1,0 +1,58 @@
+.modal {
+  @include flex(row nowrap);
+  background-color: $transparentBlack;
+  position: absolute;
+  top: 0; //updates in react editTodo and modal component
+  left: 0;
+  width: 100vw;
+  height: 100vh;
+  .modal-container {
+    @include border;
+    @include flex(column nowrap);
+    border-radius: 10px;
+    height: 30vh;
+    width: 40vw;
+    min-width: 200px;
+    min-height: 100px;
+    position: absolute;
+    padding: 1rem;
+    z-index: 1000;
+    background: var(--white, $white);
+    .modal-text {
+      width: 100%;
+      height: 70%;
+      font-size: 1.2rem;
+      color: var(--black, $black);
+    }
+    .modal-btn {
+      width: 100%;
+      height: 30%;
+      display: flex;
+      flex-flow: row nowrap;
+      justify-content: space-between;
+      align-items: center;
+      button {
+        border: none;
+        width: 40%;
+        height: 50px;
+        border-radius: 5px;
+        color: var(--black, $black);
+        font-size: 1.5rem;
+        transition: transform 1s;
+        cursor: pointer;
+      }
+      button:hover {
+        transform: scale(1.1);
+      }
+      button:active {
+        transform: scale(0.5);
+      }
+      .yes {
+        background: $red;
+      }
+      .no {
+        background: $green;
+      }
+    }
+  }
+}
